refactor: extract xdebug fucntions to dedicate file
